Mark CodeDomLocalizationProvider as partial and move classes contained in CodeDomLocalizationProvider to their own files
